libs/dev-cluster/pyproject.toml
[project]
name = "dev-cluster"
version = "0.1.0"
description = "Local development cluster orchestration"
requires-python = ">=3.11"
dependencies = ["click==8.1.8", "pyyaml==6.0.2"]

[project.scripts]
dev-cluster = "dev_cluster.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dev_cluster"]
